requirements.txt
asgiref==3.2.5
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
mysql-client==0.0.1
mysqlclient==1.4.6
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
